Use flavour-agnostic template file names
The *.in template files currently need to be renamed when switching
package flavours. To facilitate package flavour switching, we use
the flavour-agnostic templates nvidia-graphics-drivers-*.in in the
debian/templates directory and generate flavour specific files
from them.

After this change, the procedure to switch flavours is much simpler:

1) Update debian/changelog with the appropriate source package name
2) `make -f debian/rules regen-from-templates`

This work is based on the original implementation by
Daniel Dadap <ddadap@nvidia.com>.