Avoid extracting unnecessary .run files
The 64-bit package needs both the 32- and 64-bit .run files, but the 32-bit package does not. Change the unpackaging rules to unpack only the native package by default, and only unpack the 32-bit x86 package on amd64.

Also, instead of changing the .run file permissions to ensure that the .run file is executable, just call it as an argument to sh.

Change-Id: Iac3c5c678ce199c63e0f34c2e393637de6e697f4