Use the libglvnd versions of libGL.so and libEGL.so.
Switch to installing and using the libglvnd libraries for i386 and x86-64,
instead of installing both and using the non-libglvnd ones.

Change the symlinks for libGL.so and libEGL.so so that they point to the
libglvnd libraries.

Remove the non-libglvnd versions of libGL.so and libEGL.so.

Update the Vulkan ICD file to point to libGLX_nvidia.so.0 instead of
libGL.so.1.