Allow nvidia-powerd to fail
nvidia-powerd exits with an error code when the hardware is unsupported.
That marks the unit as failed, which is not ideal and breaks certification.

The nvidia 570+ driver exposes a procfs entry with the supported state of
Dynamic Boost, but it is unreliable as sometimes it reports an unknown state
which makes it hard to reliably use as a precondition.

Let's just allow the program to fail without marking the unit as such.

Co-authored-by: Kuba Pawlak <kuba.pawlak@canonical.com>