Add Conflicts with libnvidia-egl-gbm1
Nvidia provides the libnvidia-egl-* libraries both as blobs and as FOSS.
We ship libnvidia-egl-wayland.so.1 as a source-built package and Depends on it,
but have been shipping libnvidia-egl-gbm.so.1 as a blob in libnvidia-gl, so add
an explicit Conflicts statement because both packages provide the same library.

There are two other libraries libnvidia-egl-{xcb,xlib}.so.1, but those do not
(yet?) have packages in Debian/Ubuntu.

LP: #2070598